+++ CHANGELOG.md
@@ -23,6 +23,8 @@ Usage:
 ### 1.12-SNAPSHOT
 * Fix #1179: Move storageClass related functionality out of VolumePermissionEnricher to PersistentVolumeClaimStorageClassEnricher
 * Fix #1273: Deprecate `jkube.io` annotation prefix in favor of `jkube.eclipse.org` for JKubeAnnotations
+* Fix #1929: Docker Image Name parsing fix
+* Fix #2079: Bump kubernetes-client to 6.5.0
 * Fix #2086: Allow concurrent remote-dev sessions (service selector includes session id)
 * Fix #2093: ClassCastException when extracting plugins from pom
 * Fix #2104: Bump kubernetes-client to 6.5.1 (#2079)